Initial commit
analyze_if utilizes arguments and if statements to analyze one aspect of the log file at a time and produce graphs as well as new log outputs as requested.
analyze_dev is the main development file for further evolution of the analysis script. Major changes and updated should be placed here first.
formamide2.log provides the template/example necessary for the scripts to analyze.
log and logplot... are older, extraneous files uploaded to help provide history prior to git history.